// src/api-client/models/SnowflakeInfo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Quatt API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
/**
 *
 * @export
 * @interface SnowflakeInfo
 */
export interface SnowflakeInfo {
  /**
   * Field name from Snowflake table (UPPERCASE)
   * @type {string}
   * @memberof SnowflakeInfo
   */
  key: string;
  /**
   * Field value from Snowflake table (can be string, number, boolean, object, or null)
   * @type {object}
   * @memberof SnowflakeInfo
   */
  value: object;
}

/**
 * Check if a given object implements the SnowflakeInfo interface.
 */
export function instanceOfSnowflakeInfo(value: object): boolean {
  let isInstance = true;
  isInstance = isInstance && "key" in value;
  isInstance = isInstance && "value" in value;

  return isInstance;
}

export function SnowflakeInfoFromJSON(json: any): SnowflakeInfo {
  return SnowflakeInfoFromJSONTyped(json, false);
}

export function SnowflakeInfoFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): SnowflakeInfo {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    key: json["key"],
    value: json["value"],
  };
}

export function SnowflakeInfoToJSON(value?: SnowflakeInfo | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    key: value.key,
    value: value.value,
  };
}